% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTRLseq_main.R
\name{CTRLseq}
\alias{CTRLseq}
\title{CTRLseq main model}
\usage{
CTRLseq(counts, group, n.factor = 2, beta = 0.1, seed = 123)
}
\arguments{
\item{counts}{gene x sample count matrix}

\item{group}{factor (control/case)}

\item{n.factor}{number of latent factors}

\item{beta}{KL divergence weight}

\item{seed}{random seed}
}
\value{
list containing fitted model, latent factors, and DGE object
}
\description{
CTRLseq main model
}
