% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUCseq2_main.R
\name{CUCseq2}
\alias{CUCseq2}
\title{CUCseq2 main model}
\usage{
CUCseq2(counts, group, n.factor = 2, beta = 0.1, seed = 123)
}
\arguments{
\item{counts}{gene x sample count matrix}

\item{group}{factor (control/case)}

\item{n.factor}{number of latent factors}

\item{beta}{KL divergence weight}

\item{seed}{random seed}
}
\value{
list containing fitted model, latent factors, and DGE object
}
\description{
CUCseq2 main model
}
