% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEGs_CTRL}
\alias{DEGs_CTRL}
\title{Extract DEGs from CTRLseq result}
\usage{
DEGs_CTRL(obj)
}
\arguments{
\item{obj}{output from CTRLseq}
}
\value{
DEG result table
}
\description{
Extract DEGs from CTRLseq result
}
