% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEGs_CUC}
\alias{DEGs_CUC}
\title{Extract DEGs from CUCseq2 result}
\usage{
DEGs_CUC(obj)
}
\arguments{
\item{obj}{output from CUCseq2}
}
\value{
DEG result table
}
\description{
Extract DEGs from CUCseq2 result
}
