% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{devianceResiduals}
\alias{devianceResiduals}
\title{Deviance residuals for NB/Poisson GLM}
\usage{
devianceResiduals(fit)
}
\arguments{
\item{fit}{edgeR glmFit object}
}
\value{
residual matrix
}
\description{
Deviance residuals for NB/Poisson GLM
}
