% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_factor.R
\name{estimate_latent_factors}
\alias{estimate_latent_factors}
\title{Estimate latent factors using VAE}
\usage{
estimate_latent_factors(
  E,
  k = 2,
  epochs = 150,
  batch_size = 32,
  beta = 0.1,
  seed = 123
)
}
\arguments{
\item{E}{residual matrix (gene x sample)}

\item{k}{number of latent factors}

\item{epochs}{training epochs}

\item{batch_size}{batch size}

\item{beta}{KL divergence weight}

\item{seed}{random seed}
}
\value{
latent factor matrix (sample x k)
}
\description{
Estimate latent factors using VAE
}
