% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CUCseq2.R
\name{run_CUCseq2}
\alias{run_CUCseq2}
\title{Run full CUCseq2 pipeline}
\usage{
run_CUCseq2(counts, group, n_factor = 2, beta = 0.1, seed = 123)
}
\arguments{
\item{counts}{gene x sample count matrix}

\item{group}{factor (control/case)}

\item{n_factor}{number of latent factors}

\item{beta}{KL divergence weight}

\item{seed}{random seed}
}
\value{
DEG result table
}
\description{
Run full CUCseq2 pipeline
}
